+++ Runtime/Scripts/RandomizedArrayExtender.cs
@@ -1,12 +1,11 @@
-﻿using UnityEngine;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
+using UnityEngine;
 
 namespace Alkibit.Collections
 {
-	public static class RandomizedArrayExtender
-	{
-		public static T GetRandomItem<T>(this T[] array) => array[Random.Range(0, array.Length)];
+    public static class RandomizedArrayExtender
+    {
+        public static T GetRandomItem<T>(this T[] array) => array[Random.Range(0, array.Length)];
         public static T GetRandomItem<T>(this List<T> array) => array[Random.Range(0, array.Count)];
     }
 }