+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## [1.0.0] 2025-04.13
+## [1.0.1] 2026.04.18
+### Refactoring and caching
+- Renamed `Alkibit.Collections.Demo` to `Alkibit.Collections.Samples`
+- New samples
+- Code refactoring
+- Caching implementation for `TargetSearch`
+- Added samples documentation to `README.md`
+- Updated `package.json` description
+- Mild fixes to `CHANGELOG.md`
+
+## [1.0.0] 2026-04.13
 ### First release
-- The first public version of the package.
+- The first public version of the package